Return correct dtype in environments

+++ src/seals/diagnostics/noisy_obs.py
@@ -62,4 +62,4 @@ class NoisyObsEnv(base_envs.ResettablePOMDP):
     def obs_from_state(self, state: np.ndarray) -> np.ndarray:
         """Returns (x, y) concatenated with Gaussian noise."""
         noise_vector = self.rand_state.randn(self._noise_length)
-        return np.concatenate([state, noise_vector])
+        return np.concatenate([state, noise_vector]).astype(np.float32)
